Phase 3-4: auth, domain layer, media storage, panorama engine
Backend
- Argon2id hashing with a decoy hash so a missing account costs the same time
  as a wrong password (no enumeration oracle)
- Opaque server-side sessions: 256-bit token, SHA-256 at rest, idle + absolute
  expiry, sessions_valid_from for bulk revocation on password change
- Token-bucket rate limiting in one atomic UPSERT so limits hold across
  instances; denied attempts are floored at -1 so an attacker cannot lock a
  victim out indefinitely
- Role/permission matrix with a single requirePermission chokepoint
- DomainError -> HTTP mapping that never leaks an internal 5xx message
- StorageDriver abstraction (local ships, s3 stubbed); media bytes never enter
  Postgres, keys are generated and path-traversal-checked
- Tour manifest builder: batched IN queries, per-field translation fallback,
  and a referential pass that drops any hotspot pointing at unpublished,
  deleted, or out-of-scope content before it reaches the browser

Panorama engine (no dependencies, 27 KB bundled vs ~600 KB for Three.js)
- WebGL equirectangular renderer; projection exposed so hotspots are real DOM
  elements (focusable, screen-reader readable) rather than canvas paint
- Preview/full dual-texture crossfade for fast first paint
- Pointer, pinch, wheel and keyboard input; inertia and auto-rotate honour
  prefers-reduced-motion

Three bugs found and fixed during verification:
- unprojectPoint applied Ry(+yaw) instead of Ry(-yaw), mirroring every
  unprojected point (caught by unit tests)
- cullFace(FRONT) discarded the whole sphere, rendering a blank screen; the
  generated winding is already front-facing from inside (caught in-browser)
- resize() ran only in the constructor, and a zero measurement from a hidden
  container shrank the buffer to 1x1; now observed and zero-guarded

30 tests pass; the engine is verified in a real browser, with DOM hotspot
positions matching painted texture markers to the exact pixel.

+++ drizzle/0000_watery_gorgon.sql
@@ -40,7 +40,7 @@ CREATE TABLE "password_reset_tokens" (
 --> statement-breakpoint
 CREATE TABLE "rate_limits" (
 	"key" text PRIMARY KEY NOT NULL,
-	"tokens" integer NOT NULL,
+	"tokens" double precision NOT NULL,
 	"updated_at" timestamp with time zone DEFAULT now() NOT NULL
 );
 --> statement-breakpoint
